HTBHF-1971 restructure guidance directory
