+++ src/web/routes/guidance/guidance.js
@@ -1,28 +1,8 @@
 const express = require('express')
-const { compose, equals, prop } = require('ramda')
 const { handleRequestForPath } = require('../application/middleware')
 const { getLanguageBase } = require('../language')
+const { getPageMetadata } = require('./get-page-meta-data')
 const { PAGES } = require('./pages')
-
-const hasMatchingPath = (path) => compose(equals(path), prop('path'))
-
-const getPreviousPage = (pages, index) => pages[index - 1]
-
-const getNextPage = (pages, index) => pages[index + 1]
-
-const getPageMetadata = (pages, path) => {
-  const pageIndexForPath = pages.findIndex(hasMatchingPath(path))
-  const page = getPageForPath(pages, path)
-
-  return {
-    activePath: path,
-    previous: getPreviousPage(pages, pageIndexForPath),
-    next: getNextPage(pages, pageIndexForPath),
-    title: page.title
-  }
-}
-
-const getPageForPath = (pages, path) => pages.find(hasMatchingPath(path))
 
 const renderGuidanceRoute = (pages, path) => (req, res) =>
   res.render(`guidance/${getLanguageBase(req.language)}/${path}`, {
@@ -44,9 +24,5 @@ const registerGuidanceRoutes = (config, steps, app) => {
 }
 
 module.exports = {
-  registerGuidanceRoutes,
-  getPageForPath,
-  getNextPage,
-  getPreviousPage,
-  getPageMetadata
+  registerGuidanceRoutes
 }
